refactor(dotnet-broker-kafka): add createConsumerServiceFile and getMessageBrokerName core functions
